requirements.txt
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
decorator==4.3.0
Django==2.0
djangorestframework==3.8.2
httplib2==0.11.3
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
oauth2client==4.1.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
django-map-widgets